ai-mvp-pt/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
pillow==10.1.0
sentencepiece==0.1.99
protobuf==3.20.3